test: See if commit saves the latest changes
The changes made to the file after "git add" will not be included
to the saved version of the file when "git commit" is issued.
However these changes are not lost either, they are still kept
in the current file in the directory. The file is marked as "changed".
Next "git add" will include this change to the next commit.

Reference#: NONE